refactor : Tradetypes card component update, trade types mobile update and Options svg icon fix
